GER-952 : Delete-project plugin : 500 server error when clicking on delete for a project
This jira was reopened as we did not process the delete request correctly whenever the database was NoteDB. The database handler for the delete-project plugin has now been updated to handle NoteDB as well as ReviewDB changes.

+++ src/main/java/com/googlesource/gerrit/plugins/deleteproject/database/DatabaseDeleteHandler.java
@@ -15,6 +15,7 @@
 package com.googlesource.gerrit.plugins.deleteproject.database;
 
 import static java.util.Collections.singleton;
+import static java.util.stream.Collectors.toList;
 
 import com.google.common.flogger.FluentLogger;
 import com.google.gerrit.extensions.registration.DynamicItem;
@@ -30,7 +31,10 @@ import com.google.gerrit.server.account.AccountState;
 import com.google.gerrit.server.account.AccountsUpdate;
 import com.google.gerrit.server.account.ProjectWatches.ProjectWatchKey;
 import com.google.gerrit.server.change.AccountPatchReviewStore;
+import com.google.gerrit.server.git.GitRepositoryManager;
 import com.google.gerrit.server.index.change.ChangeIndexer;
+import com.google.gerrit.server.notedb.ChangeNotes;
+import com.google.gerrit.server.notedb.NotesMigration;
 import com.google.gerrit.server.project.NoSuchChangeException;
 import com.google.gerrit.server.query.account.InternalAccountQuery;
 import com.google.gerrit.server.query.change.ChangeData;
@@ -51,56 +55,71 @@ import org.eclipse.jgit.errors.ConfigInvalidException;
 
 public class DatabaseDeleteHandler {
   private static final FluentLogger log = FluentLogger.forEnclosingClass();
-
   private final Provider<ReviewDb> dbProvider;
   private final StarredChangesUtil starredChangesUtil;
   private final DynamicItem<AccountPatchReviewStore> accountPatchReviewStore;
   private final ChangeIndexer indexer;
   private final Provider<InternalAccountQuery> accountQueryProvider;
-  private final Provider<InternalChangeQuery> changeQueryProvider;
   private final Provider<AccountsUpdate> accountsUpdateProvider;
+  private final ChangeNotes.Factory schemaFactoryNoteDb;
+  private final GitRepositoryManager repoManager;
+  private final NotesMigration migration;
 
   @Inject
   public DatabaseDeleteHandler(
-          Provider<ReviewDb> dbProvider,
-          StarredChangesUtil starredChangesUtil,
-          DynamicItem<AccountPatchReviewStore> accountPatchReviewStore,
-          ChangeIndexer indexer,
-          Provider<InternalAccountQuery> accountQueryProvider,
-          Provider<InternalChangeQuery> changeQueryProvider, @UserInitiated Provider<AccountsUpdate> accountsUpdateProvider) {
+      Provider<ReviewDb> dbProvider,
+      StarredChangesUtil starredChangesUtil,
+      DynamicItem<AccountPatchReviewStore> accountPatchReviewStore,
+      ChangeIndexer indexer,
+      ChangeNotes.Factory schemaFactoryNoteDb,
+      NotesMigration migration,
+      GitRepositoryManager repoManager,
+      Provider<InternalAccountQuery> accountQueryProvider,
+      @UserInitiated Provider<AccountsUpdate> accountsUpdateProvider) {
     this.dbProvider = dbProvider;
     this.starredChangesUtil = starredChangesUtil;
     this.accountPatchReviewStore = accountPatchReviewStore;
     this.indexer = indexer;
     this.accountQueryProvider = accountQueryProvider;
-    this.changeQueryProvider = changeQueryProvider;
     this.accountsUpdateProvider = accountsUpdateProvider;
+    this.schemaFactoryNoteDb = schemaFactoryNoteDb;
+    this.repoManager = repoManager;
+    this.migration = migration;
   }
 
-  public void delete(Project project) throws OrmException {
+
+  public void delete(Project project) throws OrmException, IOException {
     ReviewDb db = ReviewDbUtil.unwrapDb(dbProvider.get());
-    Connection conn = ((JdbcSchema) db).getConnection();
-    try {
-      conn.setAutoCommit(false);
+    if (isReviewDb()) {
+      Connection conn = ((JdbcSchema) db).getConnection();
       try {
-        atomicDelete(db, project, getChangesList(project, conn));
-        conn.commit();
-      } finally {
-        conn.setAutoCommit(true);
+        conn.setAutoCommit(false);
+        try {
+          atomicDelete(db, project, getChangesList(project, conn));
+          conn.commit();
+        } finally {
+          conn.setAutoCommit(true);
+        }
+      } catch (SQLException e) {
+        try {
+          conn.rollback();
+        } catch (SQLException ex) {
+          throw new OrmException(ex);
+        }
+        throw new OrmException(e);
       }
-    } catch (SQLException e) {
-      try {
-        conn.rollback();
-      } catch (SQLException ex) {
-        throw new OrmException(ex);
-      }
-      throw new OrmException(e);
+    } else {
+      atomicDelete(db, project, getChangesListFromNoteDb(project));
     }
+  }
+
+  private boolean isReviewDb() {
+    return !migration.disableChangeReviewDb();
   }
 
   private List<Change.Id> getChangesList(Project project, Connection conn) throws SQLException {
     try (PreparedStatement changesForProject =
-        conn.prepareStatement("SELECT change_id FROM changes WHERE dest_project_name = ?")) {
+             conn.prepareStatement("SELECT change_id FROM changes WHERE dest_project_name = ?")) {
       changesForProject.setString(1, project.getName());
       try (java.sql.ResultSet resultSet = changesForProject.executeQuery()) {
         List<Change.Id> changeIds = new ArrayList<>();
@@ -114,27 +133,39 @@ public class DatabaseDeleteHandler {
     }
   }
 
+  private List<Change.Id> getChangesListFromNoteDb(Project project) throws IOException {
+    Project.NameKey projectKey = project.getNameKey();
+    List<Change.Id> changeIds =
+        schemaFactoryNoteDb
+            .scan(repoManager.openRepository(projectKey), dbProvider.get(), projectKey)
+            .map(ChangeNotes.Factory.ChangeNotesResult::id)
+            .collect(toList());
+    log.atFine().log(
+        "Number of changes in noteDb related to project %s are %d",
+        projectKey.get(), changeIds.size());
+    return changeIds;
+  }
+
+
   private void deleteChanges(ReviewDb db, Project.NameKey project, List<Change.Id> changeIds)
       throws OrmException {
-
     for (Change.Id id : changeIds) {
       try {
         starredChangesUtil.unstarAll(project, id);
       } catch (NoSuchChangeException e) {
         // we can ignore the exception during delete
       }
-      ResultSet<PatchSet> patchSets = db.patchSets().byChange(id);
-      if (patchSets != null) {
-        deleteFromPatchSets(db, patchSets);
+      if (isReviewDb()) {
+        ResultSet<PatchSet> patchSets = db.patchSets().byChange(id);
+        if (patchSets != null) {
+          deleteFromPatchSets(db, patchSets);
+        }
+        // In the future, use schemaVersion to decide what to delete.
+        db.patchComments().delete(db.patchComments().byChange(id));
+        db.patchSetApprovals().delete(db.patchSetApprovals().byChange(id));
+        db.changeMessages().delete(db.changeMessages().byChange(id));
+        db.changes().deleteKeys(Collections.singleton(id));
       }
-
-      // In the future, use schemaVersion to decide what to delete.
-      db.patchComments().delete(db.patchComments().byChange(id));
-      db.patchSetApprovals().delete(db.patchSetApprovals().byChange(id));
-
-      db.changeMessages().delete(db.changeMessages().byChange(id));
-      db.changes().deleteKeys(Collections.singleton(id));
-
       // Delete from the secondary index
       try {
         indexer.delete(id);
@@ -144,34 +175,6 @@ public class DatabaseDeleteHandler {
     }
   }
 
-  private final void deleteChangeData(List<ChangeData> changeData)
-          throws OrmException {
-    ReviewDb db = ReviewDbUtil.unwrapDb(dbProvider.get());
-    for (ChangeData cd : changeData) {
-      Change.Id id = cd.getId();
-      ResultSet<PatchSet> patchSets = null;
-      patchSets = db.patchSets().byChange(id);
-      if (patchSets != null) {
-        deleteFromPatchSets(db, patchSets);
-      }
-      // In the future, use schemaVersion to decide what to delete.
-      db.patchComments().delete(db.patchComments().byChange(id));
-      db.patchSetApprovals().delete(db.patchSetApprovals().byChange(id));
-      try {
-        starredChangesUtil.unstarAll(cd.project(), id);
-      } catch (NoSuchChangeException e) {
-        // we can ignore the exception during delete
-      }
-      db.changeMessages().delete(db.changeMessages().byChange(id));
-      db.changes().delete(Collections.singleton(cd.change()));
-      // Delete from the secondary index
-      try {
-        indexer.delete(id);
-      } catch (IOException e) {
-        log.atSevere().log(String.format("Failed to delete change %s from index", id), e);
-      }
-    }
-  }
 
   private void deleteFromPatchSets(ReviewDb db, ResultSet<PatchSet> patchSets) throws OrmException {
     for (PatchSet patchSet : patchSets) {
@@ -179,6 +182,7 @@ public class DatabaseDeleteHandler {
       db.patchSets().delete(Collections.singleton(patchSet));
     }
   }
+
 
   public void atomicDelete(ReviewDb db, Project project, List<Change.Id> changeIds)
       throws OrmException {
@@ -206,28 +210,40 @@ public class DatabaseDeleteHandler {
     }
   }
 
-  public List<Change.Id> replicatedDeleteChanges(Project project) throws OrmException {
+  /**
+   * Populating the list of changes here to return to the calling method
+   * in DeleteProject as the changes list needs to be passed to the replicated call.
+   * Both ReviewDB and NoteDb changes list supported.
+   * @param project
+   * @return
+   * @throws OrmException
+   * @throws IOException
+   */
+  public List<Change.Id> getReplicatedDeleteChangeIdsList(Project project) throws OrmException, IOException {
     ReviewDb db = ReviewDbUtil.unwrapDb(dbProvider.get());
-    // http://docs.oracle.com/javase/specs/jls/se7/html/jls-14.html#jls-14.20.3.2
-    Connection conn = ((JdbcSchema) db).getConnection();
     List<Change.Id> changes;
-    try {
-      conn.setAutoCommit(false);
+    if (isReviewDb()) {
+      Connection conn = ((JdbcSchema) db).getConnection();
       try {
-        //Need to populate the list of changes here to return to the calling method
-        //in DeleteProject as the changes list needs to be passed to the replicated call.
-        changes = getChangesList(project, conn);
-        atomicDelete(db, project, changes);
-      } finally {
-        conn.setAutoCommit(true);
+        conn.setAutoCommit(false);
+        try {
+          changes = getChangesList(project, conn);
+          atomicDelete(db, project, getChangesList(project, conn));
+          conn.commit();
+        } finally {
+          conn.setAutoCommit(true);
+        }
+      } catch (SQLException e) {
+        try {
+          conn.rollback();
+        } catch (SQLException ex) {
+          throw new OrmException(ex);
+        }
+        throw new OrmException(e);
       }
-    } catch (SQLException e) {
-      try {
-        conn.rollback();
-      } catch (SQLException ex) {
-        throw new OrmException(ex);
-      }
-      throw new OrmException(e);
+    } else {
+      changes = getChangesListFromNoteDb(project);
+      atomicDelete(db, project, getChangesListFromNoteDb(project));
     }
     return changes;
   }
